Refactor: encapsulamiento de las peticiones a bases de datos
